refactor: toolbar 중복 코드 최소화
