+++ app/src/main/java/woowacourse/movie/activity/BaseActivity.kt
@@ -1,0 +1,22 @@
+package woowacourse.movie.activity
+
+import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
+
+abstract class BaseActivity : AppCompatActivity() {
+
+    open fun setBackToBefore(id: Int) {
+        val toolbar = findViewById<Toolbar>(id)
+        setSupportActionBar(toolbar)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == android.R.id.home) {
+            finish()
+            return true
+        }
+        return super.onOptionsItemSelected(item)
+    }
+}
